Separate all-gather matmul communication implementations and refresh benchmarks
